Morpion/tests/modelTests/MorpionSolitaireModelTest.java
package modelTests;

import static org.junit.Assert.*;
import org.junit.Before;
import org.junit.Test;
import model.Grid;
import model.Line;
import model.Mode;
import model.MorpionSolitaireModel;
import model.Point;

public class MorpionSolitaireModelTest {

    private MorpionSolitaireModel morpionSolitaireModel;

    @Before
    public void setUp() {
        morpionSolitaireModel = new MorpionSolitaireModel();
    }

    @Test
    public void testInitialization() {
        assertNotNull(morpionSolitaireModel);
        assertFalse(morpionSolitaireModel.isGameOver());
        assertNull(morpionSolitaireModel.getPlayerName());
    }

    @Test
    public void testInit() {
        morpionSolitaireModel.init();

        assertNotNull(morpionSolitaireModel);
        assertFalse(morpionSolitaireModel.isGameOver());
        assertNull(morpionSolitaireModel.getPlayerName());
    }

    @Test
    public void testHandleHumanMove() {
        // You may need to modify this test based on the actual implementation of handleHumanMove

        morpionSolitaireModel.handleHumanMove(0, 0);

        // Add assertions based on the expected behavior of handleHumanMove
    }

    @Test
    public void testHandleRandomMove() {
        // You may need to modify this test based on the actual implementation of handleRandomMove

        morpionSolitaireModel.handleRandomMove();

        // Add assertions based on the expected behavior of handleRandomMove
    }

    @Test
    public void testHandleRandomGame() {
        // You may need to modify this test based on the actual implementation of handleRandomGame

        morpionSolitaireModel.handleRandomGame();

        // Add assertions based on the expected behavior of handleRandomGame
    }

    @Test
    public void testHandleNmcsMove() {
        // You may need to modify this test based on the actual implementation of handleNmcsMove

        morpionSolitaireModel.handleNmcsMove();

        // Add assertions based on the expected behavior of handleNmcsMove
    }

    @Test
    public void testHandleNmcsGame() {
        // You may need to modify this test based on the actual implementation of handleNmcsGame

        morpionSolitaireModel.handleNmcsGame();

        // Add assertions based on the expected behavior of handleNmcsGame
    }

    @Test
    public void testUndoMove() {
        // You may need to modify this test based on the actual implementation of undoMove

        morpionSolitaireModel.undoMove();

        // Add assertions based on the expected behavior of undoMove
    }

    // Add more tests as needed for other methods in MorpionSolitaireModel
}

